[docs] Remove latest-version/link-to-changelog from homepage
Changelog page can be found in the 'Information' menu,
and the versions information is visible in the badges on home page.

Therefore this text with the latest version and a link to the changelog
is kind of unnecessary.

Plus it's not that easy to set-up with automated versioning because
the gradle plugin does not currently support printing latest
effective version instead of calculating a new one. Although a PR
for that exists: https://github.com/vivin/gradle-semantic-build-versioning/pull/118